thevu: fix Services

+++ src/main/java/com/example/cnpm_backend/service/CanCuocService.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CanCuocService extends JpaRepository<CanCuocModel, Long> {
+public interface CanCuocService extends JpaRepository<CanCuocModel, Integer> {
 }
